+++ cl_dll/inputw32.cpp
@@ -66,7 +66,7 @@ cvar_t		*sensitivity;
 int			mouse_buttons;
 int			mouse_oldbuttonstate;
 POINT		current_pos;
-int			mouse_x, mouse_y, old_mouse_x, old_mouse_y, mx_accum, my_accum;
+float		mouse_x, mouse_y, old_mouse_x, old_mouse_y, mx_accum, my_accum;
 
 static int	restore_spi;
 static int	originalmouseparms[3], newmouseparms[3] = {0, 0, 1};
@@ -299,7 +299,7 @@ IN_MouseMove
 */
 void IN_MouseMove ( float frametime, usercmd_t *cmd)
 {
-	int		mx, my;
+	float	mx, my;
 	vec3_t viewangles;
 
 	gEngfuncs.GetViewAngles( (float *)viewangles );
@@ -385,7 +385,7 @@ void IN_MouseMove ( float frametime, usercmd_t *cmd)
 //#define TRACE_TEST
 #if defined( TRACE_TEST )
 	{
-		int mx, my;
+		float mx, my;
 		void V_Move( int mx, int my );
 		IN_GetMousePos( &mx, &my );
 		V_Move( mx, my );
